requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
aiosqlite==0.20.0
tweepy==4.16.0
openai==1.51.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
